% Generated by roxygen2 (4.0.0): do not edit by hand
\name{alleq}
\alias{alleq}
\title{Are all values equal?}
\usage{
alleq(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{logical: remove logicals?}
}
\value{
logical: \code{TRUE} if all elements are the same
}
\description{
Are all values equal?
}
\examples{
x <- rep(1,5)
alleq(x)
alleq(letters)
# [1] FALSE
}
\author{
Mark Cowley, 2011-07-18
}

