\% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cat.matrix}
\alias{cat.matrix}
\title{Print a matrix to stdout.}
\usage{
cat.matrix(x, row.names = TRUE, col.names = TRUE, pad = TRUE, sep = " ")
}
\arguments{
\item{x}{a \code{matrix}}

\item{row.names}{logical: print the row names?}

\item{col.names}{logical: print the col names?}

\item{pad}{logical: pad the values in each column?}

\item{sep}{the separator between columns.}
}
\value{
none. it prints a matrix to console.
}
\description{
Print a matrix to stdout.
}
\author{
Mark Cowley
}

