% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cm2inch}
\alias{cm2inch}
\title{Convert centimetres to inches}
\usage{
cm2inch(cm)
}
\arguments{
\item{cm}{vector of measurements in cM}
}
\value{
vector of measurements in inches
}
\description{
Convert centimetres to inches
}
\author{
Mark Cowley, 2009-03-18
}

