% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.na.R
\name{count.na}
\alias{count.na}
\title{count the number of NA's}
\usage{
count.na(x)
}
\arguments{
\item{x}{a 1D or 2D object}
}
\value{
a numeric(1) indicating how many \code{NA}'s were present.
}
\description{
count the number of NA's
}
\author{
Mark Cowley, 31 Oct 2004
}

