\% Generated by roxygen2 (4.0.0): do not edit by hand
\name{debug2}
\alias{debug2}
\title{Debug a function, controlled by a boolean}
\usage{
debug2(fun, text = "", condition = NULL, doit = interactive())
}
\arguments{
\item{fun}{any interpreted R function. see \code{\link[base]{debug}}}

\item{text}{a text string that can be retrieved when the browser is entered.
see \code{\link[base]{debug}}}

\item{condition}{a condition that can be retrieved when the browser is entered.
see \code{\link[base]{debug}}}

\item{doit}{logical: if TRUE, then debugging mode will be entered, if \code{FALSE},
no debugging will occur. Default=\code{TRUE} if running an interactive session.}
}
\value{
nothing
}
\description{
The standard \code{\link[base]{debug}} function works great, but this function lets you turn
debuging on and off via a logical global parameter, meaning no more code
commenting
}
\author{
Mark Cowley, 2011-08-12
}
\seealso{
\code{\link[base]{debug}}
}

