\% Generated by roxygen2 (4.0.0): do not edit by hand
\name{density_upper_tri}
\alias{density_lower_tri}
\alias{density_upper_tri}
\title{density triangles}
\usage{
density_upper_tri(x, ...)

density_lower_tri(x, ...)
}
\arguments{
\item{x}{the data from which the estimate is to be computed.}

\item{\dots}{arguments passed to density}
}
\description{
Calculate the density of the values in the upper or lower triangle of x.
}
\author{
Mark Cowley, 4 May 2006
}

