\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/dir.Rdata.R
\name{dir.Rdata}
\alias{dir.Rdata}
\title{dir listing of R data objects}
\usage{
dir.Rdata(path = ".")
}
\arguments{
\item{path}{the path}
}
\value{
a character vector listing of all R data files within \code{path}
}
\description{
dir listing of R data objects
}
\author{
Mark Cowley
}

