% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortnum.R
\name{dir.sort}
\alias{dir.sort}
\title{sort files in a directory numerically}
\usage{
dir.sort(path = ".", pattern = NULL, ...)
}
\arguments{
\item{path}{the path to a directory. default=current working directory}

\item{pattern}{an optional regular expression.  Only file names which match
the regular expression will be returned.}

\item{\dots}{arguments passed to \code{\link{dir}}}
}
\value{
A character vector containing the names of the files in the
 specified directories, or \code{""} if there were no files.  If a path
 does not exist or is not a directory or is unreadable it is
 skipped, with a warning.
}
\description{
If your files contain a simple pattern, like:\cr
tmp1.out, tmp2.out, \dots, tmp10.out, \dots\cr
and you want to process them in this order:\cr
1,2,\dots,10,\dots, then use \code{dir.sort}.
\code{dir()} does an alphanumeric
sort, so you would get tmp10.out, tmp1.out, ...
}
\author{
Mark Cowley, 2005-05-03
}

