% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate.indices.R
\name{duplicate.indices}
\alias{duplicate.indices}
\title{For all of the duplicates in x, return the indices for each of these
duplicates.}
\usage{
duplicate.indices(x, is.sorted = FALSE)
}
\arguments{
\item{x}{a vector}

\item{is.sorted}{logical: ignored in this version}
}
\value{
a list with each element corresponding to a duplicate entry in x.
  The value of each element is the numerical indices into x where this
  duplicate appeared. The name of each element is the duplicate name.
}
\description{
For all of the duplicates in x, return the indices for each of these
duplicates.
}
\section{TODO}{
 sort x so that you only need to do a local search for duplicates,
not a global search.
}
\author{
Mark Cowley, 20 March 2007
}

