\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/even.R
\name{even}
\alias{even}
\title{Check whether a vector or matrix of numbers are even.}
\usage{
even(x)
}
\arguments{
\item{x}{a vector or matrix of numbers}
}
\value{
a logical vector or matrix, same size and dimensions as \code{x}
}
\description{
Credit goes to Gregory Warnes, since this is copied from \code{gtools},
so as to avoid loading an entire library for just one function
}
\author{
Gregory Warnes, Mark Cowley, 2/5/08
}
\seealso{
\code{\link{odd}}
}

