% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.numlines.R
\name{file.numlines}
\alias{file.numlines}
\alias{numlines}
\title{number of lines in a txt file}
\usage{
file.numlines(filename)

numlines(filename)
}
\arguments{
\item{filename}{path to a text file}
}
\value{
numeric(1)
}
\description{
number of lines in a txt file
}
\author{
Mark Cowley
}

