\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/lookUp.R
\name{has.names}
\alias{has.names}
\title{does the object have names?}
\usage{
has.names(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{TRUE} if \code{x} has names, \code{FALSE} otherwise
}
\description{
does the object have names?
}
\author{
Mark Cowley, 2009-07-09
}

