\name{is.alnum}
\alias{is.alnum}
\title{is alphanumeric?}
\usage{
  is.alnum(x)
}
\arguments{
  \item{x}{a \code{character} \code{vector}, or an object
  which can be coerced to a \code{character}
  \code{vector}.}
}
\value{
  logical, length 1
}
\description{
  is alphanumeric?
}
\author{
  Mark Cowley, 2012-07-16
}

