\% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is.dir}
\alias{is.dir}
\alias{is.file}
\title{Is the given path a file or directory?}
\usage{
is.dir(path)

is.file(path)
}
\arguments{
\item{path}{the path to a file or directory}
}
\value{
\code{is.dir}: \dQuote{TRUE} if the path is a directory, \dQuote{FALSE} otherwise

\code{is.file}: \dQuote{TRUE} if the path is a file, \dQuote{FALSE} otherwise
}
\description{
Is the given path a file or directory?
}
\author{
Mark Cowley
}

