\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/is.sorted.R
\name{is.sorted}
\alias{is.sorted}
\title{Test if an Object is Sorted}
\usage{
is.sorted(x, na.rm = FALSE, strictly = FALSE)
}
\arguments{
\item{x}{an \R object with a class or a numeric, complex, character,
    logical or raw vector.}

\item{na.rm}{logical.  Should missing values be removed before checking?}

\item{strictly}{logical indicating if the check should be for
    \emph{strictly} increasing values.}
}
\value{
A length-one logical value.  All objects of length 0 or 1 are
sorted: the result will be \code{NA} for objects of length 2 or more
except for atomic vectors and objects with a class (where the \code{>=}
or \code{>} method is used to compare \code{x[i]} with \code{x[i-1]} for \code{i} in
\code{2:length(x)}).
}
\description{
Test if an object \emph{is} sorted, without the cost of sorting it.
}
\note{
This function is designed for objects with one-dimensional
  indices, as described above.  \code{data.frame}, \code{matrices} and other
  \code{arrays} may give surprising results.
}
\author{
Mark Cowley, 2012-07-16
}
\seealso{
\code{\link[base]{is.unsorted}} \code{\link[base]{sort}} \code{\link[base]{order}}
}

