% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is.unique}
\alias{is.unique}
\title{are the elements in x all unique?}
\usage{
is.unique(x)
}
\arguments{
\item{x}{a vector or a data frame or an array or \code{NULL}.}
}
\description{
are the elements in x all unique?
}
\author{
Mark Cowley, 6 April 2006
}

