\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/isWindows.R
\name{isMac}
\alias{isMac}
\title{isMac}
\usage{
isMac()
}
\value{
\code{TRUE} if current platform is Mac, \code{FALSE} otherwise
}
\description{
isMac
}
\examples{
isMac()
}
\author{
Mark Cowley, 2011-10-21
}

