% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWindows.R
\name{isWindows}
\alias{isWindows}
\title{isWindows}
\usage{
isWindows()
}
\value{
\code{TRUE} if current platform is Windows, \code{FALSE} otherwise
}
\description{
isWindows
}
\examples{
isWindows()
}
\author{
Mark Cowley, 2011-10-21
}

