% Generated by roxygen2 (4.0.0): do not edit by hand
\name{keyval2hashtable}
\alias{keyval2hashtable}
\title{Convert a vector of keys and values to a hashtable
Take a vector of keys, a vector of values, and convert into a hashtable.}
\usage{
keyval2hashtable(keys, values, make.unique = TRUE, sort.method = sort)
}
\arguments{
\item{keys}{a character vector, with keys}

\item{values}{a vector of values. \code{length(values)==length(keys)}}

\item{make.unique}{logical: keep redundant key->value pairs?}

\item{sort.method}{the method to use to sort the values for each key}
}
\description{
Convert a vector of keys and values to a hashtable
Take a vector of keys, a vector of values, and convert into a hashtable.
}
\note{
Hashtable keys have to be <255 characters in length.
}
\author{
Mark Cowley, 2009-07-13
}

