% Generated by roxygen2 (4.0.0): do not edit by hand
\name{list2adjmat}
\alias{list2adjmat}
\title{Convert a list to an adjacency matrix}
\usage{
list2adjmat(x)
}
\arguments{
\item{x}{a named list, where each element is a character vector. If it's
unnamed, names will be assigned}
}
\value{
a data.frame of 0 or 1, where rows correspond to unique words,
  columns represent each list element.
}
\description{
If x is a list of variably lengthed vectors, then return a table which has
the
unique values as the rownames, and the names(x) as the colnames, with either
a 0 or 1
depending on whether the entity in row i was found in the vector in column
j. Sorted by
decreasing frequency across the list of vectors (ie row 1 contains the word
that was
found in the most number of vectors in list x)
}
\examples{
mylist <- list(A=letters[1:15], B=letters[10:26], C=letters[15:20])
list2adjmat(mylist)
}
\author{
Mark Cowley, 2009-10-29
}

