% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd.R
\name{odd}
\alias{odd}
\title{Check whether a vector or matrix of numbers are odd.}
\usage{
odd(x)
}
\arguments{
\item{x}{a vector or matrix of numbers}
}
\value{
a logical vector or matrix, same size and dimensions as \code{x}
}
\description{
Credit goes to Gregory Warnes, since this is copied from \code{gtools}, 
so as to avoid loading an entire library for just one function
}
\author{
Gregory Warnes, Mark Cowley, 2/5/08
}
\seealso{
\code{\link{even}}
}

