% Generated by roxygen2 (4.0.0): do not edit by hand
\name{padNA}
\alias{padNA}
\title{make a ragged list uniform in length}
\usage{
padNA(l)
}
\arguments{
\item{l}{a list of vectors}
}
\value{
a list of vectors of the same length
}
\description{
take a ragged list, that is a list of vectors of different lengths, and
make it into a non-ragged list, by appending \code{NA} to the end.
}
\examples{
l <- list(1:3, 1:5 ,1:10)
padNA(l)
}
\author{
Mark Cowley, 2013-09-18
}

