% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.list.R
\name{rbind.list}
\alias{rbind.list}
\title{rbind the elements in a list together.}
\usage{
rbind.list(..., sep = ".")
}
\arguments{
\item{sep}{currently unused}

\item{\dots}{either a single list of tables, or multiple tables to rbind, where table means
\code{matrix} or \code{data.frame}}
}
\value{
a large table with all elements in the list rbind-ed
}
\description{
You can either supply many variables that are row-bindable, (ie are
the right dimension), or one list argument containing matrices that
can be row-binded
}
\details{
rownames will be set to <list element name>.<original row name>

Often when rbind-ing tables of data together, you want to tag all rows 
from each table with the sample to which they belong. You can specify
\code{name2colname}, which will add a new column and track the origin
of those rows.
}
\author{
Mark Cowley, 5 April 2006
}

