\% Generated by roxygen2 (4.0.0): do not edit by hand
\name{replace.cells}
\alias{replace.cells}
\title{replace the values in a table}
\usage{
replace.cells(tab, row, col, val)
}
\arguments{
\item{tab}{a matrix or data.frame of values, to be edited}

\item{row}{a vector of row names or indices}

\item{col}{a vector of col names or indices}

\item{val}{a vector of values to insert into tab at the locations specified by row and col}
}
\value{
a \code{matrix} or \code{data.frame} like tab, with those values specified by the
\code{row}, \code{col} tuples changed to the values specified by \code{val}.
}
\description{
replace the values in a table, at specific positions, referenced by the row name/index,
the column name/index.
}
\details{
Common usage scenario\cr
Given a spreadsheet of patients with mutated genes, you want to indicate in a square matrix
of genes x patients, which ones have a mutation. Thus a priori, you know which elements
you want to edit, but it's very fiddly using the "[", "]" nomenclature to indicate which
cells within the matrix to edit.
}
\examples{
mat <- matrix(0, 5, 5)
replace.cells(mat, 1:5, 1:5, rep(NA, 5))
df <- as.data.frame(mat)
dimnames(df) <- list(letters[1:5], LETTERS[1:5])
replace.cells(df, c("a", "a", "e"), c("D", "B", "A"), 1:3)
}
\author{
Mark Cowley, 2011-09-20
}

