\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/rm.na.R
\name{rm.na}
\alias{rm.na}
\title{remove NA's from an object}
\usage{
rm.na(x)
}
\arguments{
\item{x}{a 1D object}
}
\value{
an object with no \code{NA}'s
}
\description{
remove NA's from an object
}
\note{
Defunct: \code{\link{na.omit}} is a better alternative.
}
\author{
Mark Cowley, 2012-07-16
}

