\% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rowapply}
\alias{rowapply}
\title{apply FUN to rows}
\usage{
rowapply(x, FUN, ...)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame}}

\item{FUN}{a function}

\item{\dots}{further arguments passed to FUN}
}
\value{
the result of running FUN on each row
}
\description{
apply FUN to rows
}
\author{
Mark Cowley
}

