\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/matrix.utils.R
\name{rowVar}
\alias{rowVar}
\title{Variance of each row}
\usage{
rowVar(x, na.rm = TRUE)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame}}

\item{na.rm}{logical: remove \code{NA}'s?}
}
\value{
the variance of each row
}
\description{
Variance of each row
}
\author{
Mark Cowley
}

