\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/scale2.R
\name{scale2}
\alias{scale2}
\title{Function to linearly scale a vector of numbers (x) to be between min and
max.}
\usage{
scale2(x, min, max)
}
\arguments{
\item{x}{a numeric vector}

\item{min}{the minimum of the new range}

\item{max}{the maximum of the new range}
}
\description{
Function to linearly scale a vector of numbers (x) to be between min and
max.
}
\examples{
options(digits=2)
scale2(1:10,0,1)
# [1] 0.00 0.11 0.22 0.33 0.44 0.56 0.67 0.78 0.89 1.00
}
\author{
Mark Cowley, 31 Oct 2005
}

