\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/readLine.R
\name{skipLine}
\alias{skipLine}
\title{skipLine}
\usage{
skipLine(file)
}
\arguments{
\item{file}{an open connection}
}
\value{
nothing
}
\description{
skip a line from an open file connection
}
\author{
Mark Cowley
}

