\docType{methods}
\name{sort}
\alias{sort}
\alias{sort,data.frame-method}
\alias{sort,matrix-method}
\title{sorting or ordering more complex objects}
\arguments{
  \item{FUN}{a sort function. if \code{x} is 2D or more,
  then this is applied to the 1st dimension (ie the rows)}
}
\value{
  a sorted object of same type as \code{x}
}
\description{
  sorting or ordering more complex objects
}
\author{
  Mark Cowley
}

