\name{str.left}
\alias{str.left}
\title{From a character vector of lenght >= 1, determine the first n characters.}
\usage{
  str.left(x, n)
}
\arguments{
  \item{x}{a character vector}

  \item{n}{the number of characters to be returned}
}
\description{
  If n is larger then the length of the word, the entire
  word is returned
}
\author{
  Mark Cowley, 2009-01-09
}

