\name{str.right}
\alias{str.right}
\title{From a character vector of lenght >= 1, determine the last n characters.}
\usage{
  str.right(x, n)
}
\arguments{
  \item{x}{a character vector}

  \item{n}{the number of characters to be returned}
}
\description{
  If n is larger then the length of the word, the entire
  word is returned
}
\author{
  Mark Cowley, 2009-01-09
}

