\name{str_right}
\alias{str_left}
\alias{str_right}
\title{From a character vector of length >= 1, determine the first or last n characters.}
\usage{
  str_right(x, n)

  str_left(x, n)
}
\arguments{
  \item{x}{a character vector}

  \item{n}{the number of characters to be returned}
}
\description{
  If n is larger then the length of the word, the entire
  word is returned
}
\author{
  Mark Cowley, 2009-01-09
}

