% Generated by roxygen2 (4.0.0): do not edit by hand
\name{subdirs.same}
\alias{subdirs.same}
\title{Are the subdirs that are named to same actually identical}
\usage{
subdirs.same(x, y)
}
\arguments{
\item{x}{character(1) the path to a directory}

\item{y}{character(1) the path to a directory}
}
\value{
logical(N), per subdir found in both dirs.
}
\description{
Given 2 dirs that may share a number of subdirs, which at least by
name look to be similar, this function checks whether those directories
actually are the same in terms of file size.
}
\details{
Why bother? If your data is on local and server, then invariably this data
will be out of sync, and one will be newer than the other. it's often very
difficult to determine this just by looking at the folder names.
}
\author{
Mark Cowley, 2012-02-13
}

