\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/sunique.R
\name{sunique}
\alias{sunique}
\title{sort and uniquify a vector}
\usage{
sunique(x, ...)
}
\arguments{
\item{x}{a vector}

\item{\dots}{additional arguments passed to \code{\link{unique}}}
}
\value{
a vector of unique, sorted values
}
\description{
sort and uniquify a vector
}
\examples{
sunique(c("b", "z", "a", "z", NA))
# [1] "a" "b" "z"
}
\author{
Mark Cowley, 23/3/07
}

