% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_diff.R
\name{time_diff}
\alias{time_diff}
\title{Determines the difference between two time points generated from calling
date()}
\usage{
time_diff(t1, t2, units = "min")
}
\arguments{
\item{t1}{a date string eg "Wed Oct 20 14 03:12 2004"}

\item{t2}{as for \code{t1}}

\item{units}{one of \dQuote{sec}, \dQuote{min} or \dQuote{hr}}
}
\value{
amount oftime between t2 - t1.
}
\description{
Determines the difference between two time points generated from calling
date()
}
\note{
check out lubridate package.
}
\section{TODO}{

Make the function work for when t2 and t1 are on different days.
}
\author{
Mark Cowley, 20 Oct 2004
}

