% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichN.R
\name{which.minN}
\alias{maxN}
\alias{minN}
\alias{which.maxN}
\alias{which.minN}
\title{whichN}
\usage{
which.minN(x, N = 5, return.ties = FALSE)

minN(x, N = 5, sort = TRUE, return.ties = FALSE)

which.maxN(x, N = 5, return.ties = FALSE)

maxN(x, N = 5, sort = TRUE, return.ties = FALSE)
}
\arguments{
\item{x}{a vector of numbers}

\item{N}{how many values are to be returned}

\item{return.ties}{logical: if there are multiple Nth smallest values, return all the indices?}

\item{sort}{if \code{FALSE}: return the N values in order of appearance;
else return the N values in increasing order.}
}
\value{
\code{which.minN}: the indices of the \code{N} smallest values in \code{x}. possibly more than \code{N} if
   \code{return.ties=TRUE} and the \code{N}th smallest value appears more than once.

\code{minN}: return the N smallest values from \code{x} (possibly sorted), possibly
more than N if return.ties=TRUE and the Nth smallest value appears more than
once.

\code{which.maxN}: return the indices of the N largest values in x. possibly more than
N if return.ties=TRUE and the Nth largest value appears more than once.

\code{maxN}: return the N largest values from x (possibly sorted), possibly more
than N if return.ties=TRUE and the Nth largest value appears more than once.
}
\description{
\code{which.minN}: Get the indices of the N smallest values in x

\code{minN}: Get the N smallest values from x

\code{which.maxN}: Get the indices of the N largest values in x

\code{maxN}: Get the N largest values from x
}
\author{
Mark Cowley, 24 May 2005
}

