% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm2inch.R
\name{cm2inch}
\alias{cm2inch}
\title{Convert centimetres to inches}
\usage{
cm2inch(cm)
}
\arguments{
\item{cm}{vector of measurements in cM}
}
\value{
vector of measurements in inches
}
\description{
Convert centimetres to inches
}
\author{
Mark Cowley, 2009-03-18
}

