% Generated by roxygen2 (4.0.0): do not edit by hand
\name{count.null}
\alias{count.null}
\title{count the number of NULL's}
\usage{
count.null(x)
}
\arguments{
\item{x}{a 1D or 2D object}
}
\value{
a numeric(1) indicating how many \code{NA}'s were present.
}
\description{
count the number of NULL's
}
\author{
Mark Cowley, 31 Oct 2004
}

