\name{density_upper_tri}
\alias{density_lower_tri}
\alias{density_upper_tri}
\title{density triangles}
\usage{
  density_upper_tri(x, ...)

  density_lower_tri(x, ...)
}
\arguments{
  \item{\dots}{arguments passed to density}

  \item{x}{the data from which the estimate is to be
  computed.}
}
\description{
  Calculate the density of the values in the upper or lower
  triangle of x.
}
\author{
  Mark Cowley, 4 May 2006
}

