% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rowMedian}
\alias{rowMedian}
\title{What is the median in each row?}
\usage{
rowMedian(x)
}
\arguments{
\item{x}{a matrix-like object}
}
\value{
a numeric vector of medians
}
\description{
What is the median in each row?
}
\author{
Mark Cowley, 2009-04-09
}

