\% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rowRange}
\alias{rowRange}
\title{What is the range of values in each row?}
\usage{
rowRange(x)
}
\arguments{
\item{x}{a matrix-like object}
}
\value{
a numeric vector of ranges
}
\description{
What is the range of values in each row?
}
\author{
Mark Cowley, 2009-04-09
}

