% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utils.R
\name{rowSD}
\alias{rowSD}
\title{Standard deviation of each row}
\usage{
rowSD(x, na.rm = TRUE)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame}}

\item{na.rm}{logical: remove \code{NA}'s?}
}
\value{
the SD of each row
}
\description{
Standard deviation of each row
}
\author{
Mark Cowley
}

