\% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{sort}
\alias{sort}
\alias{sort,ANY-method}
\alias{sort,data.frame-method}
\alias{sort,matrix-method}
\title{sorting or ordering more complex objects}
\usage{
sort(x, decreasing = FALSE, na.last = NA, FUN, ...)

\S4method{sort}{matrix}(x, decreasing = FALSE, na.last = NA, FUN, ...)

\S4method{sort}{data.frame}(x, decreasing = FALSE, na.last = NA, FUN, ...)

\S4method{sort}{ANY}(x, decreasing = FALSE, na.last = NA, FUN, ...)
}
\arguments{
\item{FUN}{a sort function. if \code{x} is 2D or more, then this is applied to the 1st
dimension (ie the rows)}

\item{x}{for \code{sort} an \R object with a class or a numeric,
    complex, character or logical vector.  For \code{sort.int}, a
    numeric, complex, character or logical vector, or a factor.}

\item{decreasing}{logical.  Should the sort be increasing or decreasing?
     Not available for partial sorting.}

\item{na.last}{for controlling the treatment of \code{NA}s.
    If \code{TRUE}, missing values in the data are put last; if
    \code{FALSE}, they are put first; if \code{NA}, they are removed.}

\item{...}{arguments to be passed to or from methods or (for the
    default methods and objects without a class) to \code{sort.int}.}
}
\value{
a sorted object of same type as \code{x}
}
\description{
sorting or ordering more complex objects
}
\author{
Mark Cowley
}

