% Generated by roxygen2 (4.0.0): do not edit by hand
\name{strcat}
\alias{strcat}
\title{concatenate a bunch of strings}
\usage{
strcat(...)
}
\arguments{
\item{\dots}{character vectors to be concatenated}
}
\description{
concatenate a bunch of strings
}
\examples{
strcat(letters)
# [1] "abcdefghijklmnopqrstuvwxyz"
}
\author{
Mark Cowley, 11 April 2006
}

