\name{sunique}
\alias{sunique}
\title{sort and uniquify a vector}
\usage{
  sunique(x, ...)
}
\arguments{
  \item{x}{a vector}

  \item{\dots}{additional arguments passed to
  \code{\link{unique}}}
}
\value{
  a vector of unique, sorted values
}
\description{
  sort and uniquify a vector
}
\author{
  Mark Cowley, 23/3/07
}

