\% Generated by roxygen2 (4.1.1): do not edit by hand
\% Please edit documentation in R/which.percentile.R
\name{which.quartile}
\alias{which.quartile}
\title{For a numeric vector x, return which quartile each data point is in}
\usage{
which.quartile(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
return the quartile of each data point is in, relative to the whole vector.
}
\description{
For a numeric vector x, return which quartile each data point is in
}
\author{
Mark Cowley, 2011-09-28
}
\seealso{
\code{\link{which.percentile}}
}

